Preferences: Redo the Keyboard preferences
Merge the "Keyboard" and "Advanced Keyboard" tabs into one "My Keyboard", drop
the unnecessary title within the tab, and the Cards too.
